Support both old & new codepaths in inject
